+++ src/main/java/org/server/remoteclass/controller/AdminController.java
@@ -1,14 +1,9 @@
 package org.server.remoteclass.controller;
 
 import io.swagger.annotations.ApiOperation;
-import org.server.remoteclass.dto.ResponseUserByAdminDto;
-import org.server.remoteclass.dto.ResponseUserDto;
-import org.server.remoteclass.dto.StudentDto;
-import org.server.remoteclass.exception.IdNotExistException;
-import org.server.remoteclass.jpa.UserRepository;
+import org.server.remoteclass.dto.user.ResponseUserByAdminDto;
 import org.server.remoteclass.service.AdminService;
 import org.server.remoteclass.service.StudentService;
-import org.server.remoteclass.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/admin")
