+++ src/main/java/org/server/remoteclass/controller/UserController.java
@@ -1,9 +1,8 @@
 package org.server.remoteclass.controller;
 
 import io.swagger.annotations.ApiOperation;
-import org.server.remoteclass.dto.ResponseUserByAdminDto;
-import org.server.remoteclass.dto.ResponseUserDto;
-import org.server.remoteclass.dto.UserDto;
+import org.server.remoteclass.dto.user.ResponseUserByAdminDto;
+import org.server.remoteclass.dto.user.ResponseUserDto;
 import org.server.remoteclass.jpa.UserRepository;
 import org.server.remoteclass.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
