+++ src/main/java/org/server/remoteclass/dto/lecture/LectureDto.java
@@ -1,4 +1,4 @@
-package org.server.remoteclass.dto;
+package org.server.remoteclass.dto.lecture;
 
 import lombok.*;
 import org.server.remoteclass.entity.Lecture;
