+++ src/main/java/org/server/remoteclass/dto/lecture/ResponseLectureDto.java
@@ -1,4 +1,4 @@
-package org.server.remoteclass.dto;
+package org.server.remoteclass.dto.lecture;
 
 import lombok.*;
 import org.server.remoteclass.entity.Student;
