+++ src/main/java/org/server/remoteclass/dto/order/OrderDto.java
@@ -1,19 +1,20 @@
-package org.server.remoteclass.dto;
+package org.server.remoteclass.dto.order;
 
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.server.remoteclass.constant.OrderStatus;
 import org.server.remoteclass.constant.Payment;
+import org.server.remoteclass.entity.Order;
 import org.server.remoteclass.entity.OrderLecture;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
+@Builder
 @Getter
 @Setter
-public class OrderFormDto {
-//    private Long orderId;
+public class OrderDto {
+    private Long orderId;
     private Long userId;       //주문하는 회원
     private List<OrderLecture> orderLectures;
     private OrderStatus orderStatus;
@@ -23,4 +24,18 @@ public class OrderFormDto {
 //    private String bank;
 //    private String account;
     private Payment payment;
+
+
+    public static OrderDto from(Order order){
+        if(order == null) return null;
+        return OrderDto.builder()
+                .userId(order.getUser().getUserId())
+                .orderLectures(order.getOrderLectures())
+                .orderStatus(order.getOrderStatus())
+                .orderDate(order.getOrderDate())
+                .payment(order.getPayment())
+                .build();
+    }
+
+
 }
