+++ src/main/java/org/server/remoteclass/jpa/IssuedCouponRepository.java
@@ -1,6 +1,5 @@
 package org.server.remoteclass.jpa;
 
-import org.server.remoteclass.dto.CouponDto;
 import org.server.remoteclass.entity.IssuedCoupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
