+++ src/main/java/org/server/remoteclass/jwt/TokenProvider.java
@@ -3,8 +3,7 @@ package org.server.remoteclass.jwt;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import lombok.extern.slf4j.Slf4j;
-import org.server.remoteclass.dto.TokenDto;
-import org.server.remoteclass.jpa.UserRepository;
+import org.server.remoteclass.dto.auth.TokenDto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
