+++ src/main/java/org/server/remoteclass/service/AdminServiceImpl.java
@@ -1,8 +1,7 @@
 package org.server.remoteclass.service;
 
 import org.modelmapper.ModelMapper;
-import org.server.remoteclass.dto.ResponseUserByAdminDto;
-import org.server.remoteclass.dto.ResponseUserDto;
+import org.server.remoteclass.dto.user.ResponseUserByAdminDto;
 import org.server.remoteclass.entity.User;
 import org.server.remoteclass.jpa.UserRepository;
 import org.server.remoteclass.util.BeanConfiguration;
