+++ src/main/java/org/server/remoteclass/service/AuthService.java
@@ -1,7 +1,10 @@
 package org.server.remoteclass.service;
 
-import org.server.remoteclass.dto.*;
-import org.springframework.security.core.userdetails.UserDetailsService;
+import org.server.remoteclass.dto.auth.LoginDto;
+import org.server.remoteclass.dto.auth.TokenDto;
+import org.server.remoteclass.dto.auth.TokenRequestDto;
+import org.server.remoteclass.dto.user.RequestUserDto;
+import org.server.remoteclass.dto.user.ResponseUserDto;
 
 public interface AuthService {
     ResponseUserDto signup(RequestUserDto requestUserDto);
