+++ src/main/java/org/server/remoteclass/service/CouponService.java
@@ -1,6 +1,6 @@
 package org.server.remoteclass.service;
 
-import org.server.remoteclass.dto.CouponDto;
+import org.server.remoteclass.dto.coupon.CouponDto;
 import org.server.remoteclass.exception.IdNotExistException;
 
 import java.util.List;
