+++ src/main/java/org/server/remoteclass/service/EventService.java
@@ -1,0 +1,14 @@
+package org.server.remoteclass.service;
+
+public interface EventService {
+
+    // 이벤트 전체 조회
+
+    // 이벤트 개별 상세 조회
+
+    // 이벤트 생성
+
+    // 이벤트 수정
+
+    // 이벤트 삭제
+}
