+++ src/main/java/org/server/remoteclass/service/OrderService.java
@@ -1,8 +1,7 @@
 package org.server.remoteclass.service;
 
-import org.server.remoteclass.dto.OrderDto;
-import org.server.remoteclass.dto.OrderFormDto;
-import org.server.remoteclass.dto.OrderLectureDto;
+import org.server.remoteclass.dto.order.OrderDto;
+import org.server.remoteclass.dto.order.OrderFormDto;
 import org.server.remoteclass.entity.OrderLecture;
 import org.server.remoteclass.exception.IdNotExistException;
