+++ src/main/java/org/server/remoteclass/service/OrderServiceImpl.java
@@ -3,9 +3,8 @@ package org.server.remoteclass.service;
 import org.modelmapper.ModelMapper;
 import org.server.remoteclass.constant.Authority;
 import org.server.remoteclass.constant.OrderStatus;
-import org.server.remoteclass.dto.OrderDto;
-import org.server.remoteclass.dto.OrderFormDto;
-import org.server.remoteclass.dto.OrderLectureDto;
+import org.server.remoteclass.dto.order.OrderDto;
+import org.server.remoteclass.dto.order.OrderFormDto;
 import org.server.remoteclass.entity.Order;
 import org.server.remoteclass.entity.OrderLecture;
 import org.server.remoteclass.entity.User;
