+++ src/main/java/org/server/remoteclass/service/StudentService.java
@@ -1,6 +1,9 @@
 package org.server.remoteclass.service;
 
-import org.server.remoteclass.dto.*;
+import org.server.remoteclass.dto.lecture.ResponseLectureDto;
+import org.server.remoteclass.dto.student.RequestStudentDto;
+import org.server.remoteclass.dto.student.ResponseStudentByLecturerDto;
+import org.server.remoteclass.dto.student.StudentDto;
 import org.server.remoteclass.exception.ForbiddenException;
 import org.server.remoteclass.exception.IdNotExistException;
 import org.server.remoteclass.exception.NameDuplicateException;
