+++ src/main/java/org/server/remoteclass/service/UserService.java
@@ -1,11 +1,8 @@
 package org.server.remoteclass.service;
-import org.server.remoteclass.dto.ResponseUserByAdminDto;
-import org.server.remoteclass.dto.ResponseUserDto;
-import org.server.remoteclass.dto.UserDto;
-import org.server.remoteclass.entity.User;
+import org.server.remoteclass.dto.user.ResponseUserByAdminDto;
+import org.server.remoteclass.dto.user.ResponseUserDto;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface UserService {
     ResponseUserDto getUserWithAuthorities(String email);
