+++ src/test/java/org/server/remoteclass/service/AuthServiceTest.java
@@ -5,10 +5,9 @@ import com.jayway.jsonpath.JsonPath;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.server.remoteclass.dto.LoginDto;
-import org.server.remoteclass.dto.RequestUserDto;
-import org.server.remoteclass.dto.TokenRequestDto;
-import org.server.remoteclass.dto.UserDto;
+import org.server.remoteclass.dto.auth.LoginDto;
+import org.server.remoteclass.dto.user.RequestUserDto;
+import org.server.remoteclass.dto.auth.TokenRequestDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
