+++ src/test/java/org/server/remoteclass/service/CouponServiceTest.java
@@ -5,7 +5,10 @@ import com.jayway.jsonpath.JsonPath;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.server.remoteclass.dto.*;
+import org.server.remoteclass.dto.auth.LoginDto;
+import org.server.remoteclass.dto.coupon.CouponDto;
+import org.server.remoteclass.dto.auth.TokenRequestDto;
+import org.server.remoteclass.dto.user.RequestUserDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
